Pages structure refacftorization
From now on, each page that will be added inside (dashboard) will need to have their own layout.tsx file too.
