+++ app/(pages)/(dashboard)/_components/CarouselCollections/CarouselCollectionsImage.tsx
