+++ app/(pages)/(dashboard)/_components/CarouselFlashSale/CarouselFlashCard.tsx
