+++ app/(pages)/(dashboard)/_components/CarouselFlashSale/CarouselFlashSale.tsx
