+++ app/(pages)/(dashboard)/_components/CarouselHero/CarouselHero.tsx
