+++ app/(pages)/(dashboard)/_components/CarouselHero/CarouselHeroImage.tsx
