+++ app/(pages)/(dashboard)/_components/CarouselShopCategory/CarouselShopCategoryImage.tsx
