+++ app/(pages)/(dashboard)/_components/NewArrivals/NewArrivalsSection.tsx
