+++ app/(pages)/(dashboard)/_components/SingIn/ForgotPasswordForm.tsx
@@ -14,7 +14,7 @@ import {
 } from "@/components/Form";
 import { Button } from "@/components/ui/Button/Button";
 import { Input } from "@/components/ui/Input";
-import { useSignInContext } from "./SignUpContext";
+import { useSignInContext } from "./SignInContext";
 
 import HeadLogo from "./HeadLogo";
