+++ app/(pages)/(dashboard)/_components/SingIn/SignIn.tsx
@@ -1,7 +1,7 @@
 import ForgotPasswordForm from "./ForgotPasswordForm";
 import LoginForm from "./LoginForm";
 import RegisterForm from "./RegisterForm";
-import { useSignInContext } from "./SignUpContext";
+import { useSignInContext } from "./SignInContext";
 
 const SignIn = () => {
   const { form } = useSignInContext();
