+++ app/(pages)/(dashboard)/_components/SingIn/SignInModal.tsx
@@ -10,7 +10,7 @@ import {
 import NiceModal, { useModal } from "@ebay/nice-modal-react";
 import { IoIosClose } from "react-icons/io";
 import SignIn from "./SignIn";
-import { useSignInContext } from "./SignUpContext";
+import { SignInProvider, useSignInContext } from "./SignInContext";
 
 export const SignInModal = NiceModal.create(() => {
   const modal = useModal();
@@ -38,7 +38,9 @@ export const SignInModal = NiceModal.create(() => {
             <DialogDescription></DialogDescription>
           </DialogHeader>
           <div className="w-full px-5 py-5 mx-auto overflow-hidden bg-white border border-gray-300 rounded-lg sm:w-96 md:w-[450px] sm:px-8">
+            <SignInProvider>
               <SignIn />
+            </SignInProvider>
           </div>
         </div>
       </DialogContent>
