+++ app/(pages)/(dashboard)/page.tsx
@@ -1,5 +1,6 @@
 import DashboardPageView from "./_components/DashboardPageView";
 
+
 export default function DashboardPage() {
   return <DashboardPageView />;
 }
