+++ app/page.tsx
@@ -1,4 +1,4 @@
-import DashboardPage from "./(pages)/dashboard/page";
+import DashboardPage from "./(pages)/(dashboard)/page";
 
 export default function Home() {
   return (
